matrics_addons/compliance_management/resources/scripts/requirements.txt
odoo_connect
click-odoo
pandas
numpy
python-dotenv